[Attention] Add out dtype for QK matmul (#1638)
This PR adds an argument `qk_dtype` for `attention.py`, essentially making `matmul(q, k, out_dtype=qk_dtype)`.

The motivation is that some models (e.g. phi), when operating under float16, can observe INF in Q@K. Thus, we accumulate in float32 to avoid this.

+++ python/mlc_chat/model/phi/phi_model.py
@@ -178,7 +178,7 @@ class PhiCrossAttention(nn.Module):
         super().__init__()
 
     def forward(self, q: Tensor, k: Tensor, v: Tensor, attention_mask: Tensor):
-        output = op_ext.attention(q, k, v, casual_mask=attention_mask)
+        output = op_ext.attention(q, k, v, casual_mask=attention_mask, qk_dtype="float32")
         return output
